moved some models to shared project
