+++ HappyTravel.Edo.Common/Models/Notifications/EmailSettings.cs
@@ -1,0 +1,9 @@
+namespace HappyTravel.Edo.Common.Models.Notifications
+{
+    public readonly struct EmailSettings
+    {
+        public string Email { get; init; }
+        public string TemplateId { get; init; }
+        public object Data { get; init; }
+    }
+}
